refactor: migrate simulation coordinates to integer game units (#37)
Simulation positions and spatial distances (radii, ranges, offsets, pushback) are now integer game units at 1000 units per tile, centralized in GameUnits. The arena spans 18000 by 32000 units; grid indices, time, damage, angles and multipliers keep their units.

- Position stores whole units with a fixed-point carry so fractional per-tick movement keeps its effective speed; clamping only touches out-of-bounds axes
- Range, sight, radius and collision checks use exact long squared distances
- Loaders convert each JSON field once based on its current unit (tile fields via tiles(), raw speeds to units per second, raw pushback passed through); resource JSON is unchanged
- Formation offsets are converted once; the circular fallback keeps its layout, with the 355 divisor renamed LEGACY_SUMMON_RADIUS_DIVISOR
- Per-tick arrival checks allow Position.MAX_ROUNDING_DISTANCE so rounding does not delay dashes, hooks or projectile hits by a tick
- Python bridge protocol and desktop rendering/input stay in tiles through boundary adapters
- Tests migrated to the same gameplay situations, plus new tests for conversions, movement, boundaries, overflow, formations and interface conversions

+++ core/src/main/java/org/crforge/core/ability/HookAbility.java
@@ -4,16 +4,16 @@ package org.crforge.core.ability;
  * Hook ability data (Fisherman). Troop throws a hook to pull the target toward itself, then drags
  * itself toward the target.
  *
- * @param hookRange maximum range for hook to trigger
- * @param hookMinimumRange minimum range for hook to trigger
+ * @param hookRange maximum edge-to-edge range for hook to trigger, in game units
+ * @param hookMinimumRange minimum edge-to-edge range for hook to trigger, in game units
  * @param hookLoadTime wind-up time before the hook fires
  * @param hookDragBackSpeed speed at which the target is pulled (raw units, divided by 60 at
  *     runtime)
  * @param hookDragSelfSpeed speed at which the fisherman pulls itself toward the target
  */
 public record HookAbility(
-    float hookRange,
-    float hookMinimumRange,
+    int hookRange,
+    int hookMinimumRange,
     float hookLoadTime,
     float hookDragBackSpeed,
     float hookDragSelfSpeed)
