+++ core/src/main/java/org/crforge/core/engine/AoeDamageEvent.java
@@ -4,6 +4,6 @@ import org.crforge.core.player.Team;
 
 /**
  * Records an instantaneous AOE damage burst for visualization. Captured in
- * CombatSystem.applySpellDamage() and consumed by the renderer.
+ * CombatSystem.applySpellDamage() and consumed by the renderer. Center and radius are game units.
  */
-public record AoeDamageEvent(float centerX, float centerY, float radius, Team sourceTeam) {}
+public record AoeDamageEvent(int centerX, int centerY, int radius, Team sourceTeam) {}
